Refactor diff view sidebar actions
